Toggle Menu and Logo
Added a toggle option for the eventual dark mode display option and changed the positioning of some elements on the page. Main focus was on the about page and will spread changes to the rest of the pages when further along in the styling process. Logo added for a temporary placeholder.